move files, fix errors
