+++ libs/sdk/testing/private/src/a11y/a11y-analyzer-config.ts
