+++ libs/sdk/testing/private/src/index.ts
@@ -1,4 +1,4 @@
-export { _SkyA11yAnalyzer } from './a11y-analyzer';
-export type { _SkyA11yAnalyzerConfig } from './a11y-analyzer-config';
+export { _SkyA11yAnalyzer } from './a11y/a11y-analyzer';
+export type { _SkyA11yAnalyzerConfig } from './a11y/a11y-analyzer-config';
 export { _skyTestingCheckVisibility } from './utility/check-visibility';
 export type { _SkyTestingCheckVisibilityOptions } from './utility/check-visibility-options';
