+++ libs/sdk/testing/vitest/src/matchers/to-be-accessible-options.ts
@@ -4,4 +4,5 @@ import { type _SkyA11yAnalyzerConfig } from '@skyux-sdk/testing/private';
 /**
  * The options for the `toBeAccessible` vitest matcher.
  */
+// eslint-disable-next-line @typescript-eslint/no-empty-object-type, @typescript-eslint/no-empty-interface
 export interface SkyToBeAccessibleOptions extends _SkyA11yAnalyzerConfig {}
