+++ libs/sdk/testing/vitest/src/matchers/to-be-visible-options.ts
@@ -4,4 +4,5 @@ import { type _SkyTestingCheckVisibilityOptions } from '@skyux-sdk/testing/priva
 /**
  * The options for the `toBeVisible` vitest matcher.
  */
+// eslint-disable-next-line @typescript-eslint/no-empty-object-type, @typescript-eslint/no-empty-interface
 export interface SkyToBeVisibleOptions extends _SkyTestingCheckVisibilityOptions {}
